+++ src/components/qualification/Qualification.jsx
@@ -1,0 +1,103 @@
+import dynamic from 'next/dynamic';
+import { useState, useEffect } from 'react';
+import getData from "@/components/_ServerHelper";
+import styles from '@/app/_root.module.css'
+
+const QualificationData = dynamic(() => import("./Data"))
+const QualificationSkeleton = dynamic(() => import("./Skeleton"))
+const ErrorFetch = dynamic(() => import("@/components/ErrorFetch"))
+
+export default function QualificationSection({ initdata }) {
+    const [data, setData] = useState(null);
+    const [loading, setLoading] = useState(true);
+    const [error, setError] = useState(false);
+    const [activeTab, setActiveTab] = useState(0);
+
+    useEffect(() => {
+        setData(initdata && initdata);
+        setError(!initdata && true);
+        setLoading(false);
+    }, [initdata])
+
+    const toggleLoading = () => {
+        setLoading((current) => (current === true ? false : true));
+    };
+
+    const toggleError = () => {
+        setError((current) => (current === true ? false : true));
+    };
+
+    const clickReload = () => {
+        setLoading(true); setError(false); setData(null);
+        const localData = localStorage.getItem('_data');
+
+        if (localData) {
+            try {
+                const dataJson = JSON.parse(localData).qualification;
+                if (dataJson) {
+                    setData(dataJson);
+                    setLoading(false);
+                    return;
+                }
+                throw new Error('Local qualification data not found');
+            } catch (error) {
+                console.error('Failed load local qualification data');
+                fetchData();
+            }
+        } else {
+            fetchData()
+        }
+    }
+
+    const clickTab = (tabIndex) => {
+        setActiveTab(tabIndex);
+    }
+
+    const fetchData = async () => {
+        try {
+            const result = await getData();
+            if (result) {
+                setData(result.qualification);
+                localStorage.setItem(`_data`, JSON.stringify(result))
+            }
+        } catch (error) {
+            console.error("Failed to fetch new data from server");
+            setError(true);
+        } finally {
+            setLoading(false);
+        }
+    }
+
+    return (
+        <section className={`${styles.qualification} ${styles.section}`} id="qualification">
+            <h2 onClick={toggleLoading} className={styles.section__title}>Qualification</h2>
+            <span onClick={toggleError} className={styles.section__subtitle}>My Personal Journey</span>
+
+            <div className={`${styles.qualification__container} ${styles.container}`}>
+                {error && <ErrorFetch clickEvent={clickReload} type={'qualification'} />}
+                <div className={styles.qualification__tabs}>
+                    {loading && <QualificationSkeleton part={'tabs'} loadingClick={toggleLoading} />}
+                    {data && !loading && !error &&
+                        <QualificationData
+                            data={data}
+                            part="tabs"
+                            tab={activeTab}
+                            tabclick={clickTab}
+                        />
+                    }
+                </div>
+
+                <div className={styles.qualification__sections}>
+                    {loading && <QualificationSkeleton part={'content'} loadingClick={toggleLoading} />}
+                    {data && !loading && !error &&
+                        <QualificationData
+                            data={data}
+                            part="content"
+                            tab={activeTab}
+                        />
+                    }
+                </div>
+            </div>
+        </section>
+    )
+}
